+++ library/src/main/java/org/mustangproject/ZUGFeRD/IExportableTransaction.java
@@ -38,6 +38,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.mustangproject.FileAttachment;
 import org.mustangproject.IncludedNote;
 import org.mustangproject.ReferencedDocument;
+import org.mustangproject.TradeParty;
 import org.mustangproject.ZUGFeRD.model.DocumentCodeTypeConstants;
 
 /***
@@ -211,6 +212,15 @@ public interface IExportableTransaction {
 	 * @return the sender's identification
 	 */
 	default String getOwnForeignOrganisationID() {
+		return null;
+	}
+
+	/**
+	 * BT-17 tender or lot reference
+	 *
+	 * @return mandatory ID, optional Date
+	 */
+	default IReferencedDocument getTenderReferencedDocument() {
 		return null;
 	}
 
@@ -493,9 +503,27 @@ public interface IExportableTransaction {
 	/***
 	 * payee / payment receiver, if different from seller, ram:Payee (only supported for zf2)
 	 *
-	 * @return the IZUGFeRDExportableTradeParty payment receiver, if different from sellver
+	 * @return the IZUGFeRDExportableTradeParty payment receiver, if different from seller
 	 */
 	default IZUGFeRDExportableTradeParty getPayee() {
+		return null;
+	}
+
+	/***
+	 * invoicer / invoice sender, if different from seller, ram:InvoicerTradeParty
+	 *
+	 * @return the IZUGFeRDExportableTradeParty invoice sender, if different from seller
+	 */
+	default IZUGFeRDExportableTradeParty getInvoicer() {
+		return null;
+	}
+
+	/***
+	 * invoicee / invoice receiver, if different from buyer, ram:InvoiceeTradeParty
+	 *
+	 * @return the IZUGFeRDExportableTradeParty invoice receiver, if different from buyer
+	 */
+	default IZUGFeRDExportableTradeParty getInvoicee() {
 		return null;
 	}
 
@@ -531,6 +559,26 @@ public interface IExportableTransaction {
 	}
 
 	default String getDespatchAdviceReferencedDocumentID() {
+		return null;
+	}
+
+	/**
+	 * get delivery note document ID
+	 * ram:ApplicableHeaderTradeDelivery/ram:DeliveryNoteReferencedDocument/IssuerAssignedID
+	 *
+	 * @return the ID of the delivery note document
+	 */
+	default String getDeliveryNoteReferencedDocumentID() {
+		return null;
+	}
+
+	/**
+	 * get delivery note document date
+	 * ram:ApplicableHeaderTradeDelivery/ram:DeliveryNoteReferencedDocument/FormattedIssueDateTime
+	 *
+	 * @return the date of the delivery note document
+	 */
+	default Date getDeliveryNoteReferencedDocumentDate() {
 		return null;
 	}